+++ CHANGES.rst
@@ -1,18 +1,25 @@
 Changelog
 =========
+0.2.2 (2015-02-02)
+------------------
 
-0.2.1 (2015-01-01)
+- now unpacks unified installer cache to ~/.buildout/
+- creates ~/.buildout/default.cfg file
+- keeps a cache based on settings in default.cfg, 
+ cache is, by default ~/.buildout/[eggs|downloads|extends-cache]
+
+0.2.1 (2015-02-01)
 ------------------
 
 - fix broken release
 
-0.2.0 (2015-01-01)
+0.2.0 (2015-02-01)
 ------------------
 
 - added and documented quick install script
 - added Change log
 
-0.1.0 (2015-01-01)
+0.1.0 (2015-02-01)
 ------------------
 
 - Initial release